Moved Slug type into types.py
